+++ src/components/Shell/AppShell.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts">
+import { useShellMenu } from '@/app/shell/menu/use'
+
+useShellMenu()
+</script>
+
+<template>
+  <slot />
+</template>
