+++ src/components/settings/storage/StorageSettingsPanel.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
 import { computed, onMounted, ref, watch } from 'vue'
-import { useClipboard } from '@vueuse/core'
 import { useRouter } from 'vue-router'
 import { useI18n } from '@open-pencil/vue'
 
@@ -13,20 +12,16 @@ import {
   storageProviderRegistry,
   writeStoragePreference
 } from '@/app/integrations/storage'
-import {
-  buildCORSConfigurationJSON,
-  collectCloudCORSOrigins
-} from '@/app/integrations/storage/s3/cors'
 import { appCredentialServices } from '@/app/settings/credentials/app'
 import { settingsDialogOpen } from '@/app/settings/dialog'
 import { credentialRef } from '@/app/settings/credentials/reference'
 import type { CredentialStatus } from '@/app/settings/credentials/types'
+import { toast } from '@/app/shell/ui'
 import { resumeStorageSync } from '@/app/storage/sync'
 import AppInput from '@/components/ui/AppInput.vue'
 
 const { dialogs } = useI18n()
 const router = useRouter()
-const { copy, copied } = useClipboard()
 const provider = computed(() => storageProviderRegistry.get(activeStorageProviderID.value))
 const preferenceDrafts = ref<Record<string, string>>({
   ...readStoragePreferences(provider.value.id)
@@ -34,7 +29,6 @@ const preferenceDrafts = ref<Record<string, string>>({
 const credentialDrafts = ref<Record<string, string>>({})
 const credentialStatuses = ref<Record<string, CredentialStatus>>({})
 const busy = ref(false)
-const result = ref<{ ok: boolean; message: string } | null>(null)
 const configured = computed(
   () =>
     storagePreferencesComplete(provider.value.id) &&
@@ -87,25 +81,19 @@ async function openWorkspace(): Promise<void> {
   await router.push('/storage')
 }
 
-function copyCORSConfiguration(): void {
-  void copy(buildCORSConfigurationJSON(collectCloudCORSOrigins()))
-}
-
 async function testConnection(): Promise<void> {
   busy.value = true
-  result.value = null
   try {
     savePreferences()
     for (const field of provider.value.credentialFields) {
       await saveCredential(field.id)
     }
     await resumeStorageSync()
-    result.value = await createActiveStorageAdapter(provider.value.id).testConnection()
+    const connection = await createActiveStorageAdapter(provider.value.id).testConnection()
+    if (connection.ok) toast.info(connection.message)
+    else toast.error(connection.message)
   } catch (error) {
-    result.value = {
-      ok: false,
-      message: error instanceof Error ? error.message : String(error)
-    }
+    toast.error(error instanceof Error ? error.message : String(error))
   } finally {
     busy.value = false
   }
@@ -114,7 +102,6 @@ async function testConnection(): Promise<void> {
 watch(activeStorageProviderID, (providerID) => {
   preferenceDrafts.value = { ...readStoragePreferences(providerID) }
   credentialDrafts.value = {}
-  result.value = null
   void refreshStatuses()
 })
 
@@ -198,15 +185,6 @@ onMounted(() => void refreshStatuses())
     </button>
 
     <button
-      v-if="provider.id === 's3-compatible'"
-      type="button"
-      class="rounded px-3 py-1.5 text-[11px] text-muted hover:bg-hover hover:text-surface"
-      @click="copyCORSConfiguration"
-    >
-      {{ copied ? dialogs.copied : dialogs.copyStorageCors }}
-    </button>
-
-    <button
       type="button"
       class="rounded border border-border px-3 py-1.5 text-[11px] font-medium text-surface hover:bg-hover disabled:text-muted disabled:opacity-50"
       :disabled="!configured"
@@ -215,14 +193,5 @@ onMounted(() => void refreshStatuses())
     >
       {{ dialogs.openStorageWorkspace }}
     </button>
-
-    <p
-      v-if="result"
-      class="rounded border border-border bg-panel px-2 py-1.5 text-[10px] text-muted data-[state=success]:text-success data-[state=error]:text-danger"
-      :data-state="result.ok ? 'success' : 'error'"
-      role="status"
-    >
-      {{ result.message }}
-    </p>
   </section>
 </template>
